+++ lighthub/abstractin.h
@@ -10,7 +10,7 @@ public:
 
 protected:
    Input * in;
-int publish(int value, char* subtopic = NULL);
+int publish(long value, char* subtopic = NULL);
 int publish(float value, char* subtopic = NULL );
 int publish(char * value, char* subtopic = NULL);
 friend Input;
